Created AsteroidSystem
Goal: Creating an object 'AsteroidSystem' which again creates an
ArrayList with Asteroids. The idea was to tidy up the code in the main
file and to be able to add and delete Asteroid objects via the
AsteroidSystem.
Result: It worked. Kind of. I can not control the amount of asteroids
created, it seems to continuously produce asteroids and slows down the
whole program.